+++ src/main/kotlin/lgp/core/environment/dataset/CsvDatasetLoader.kt
@@ -2,7 +2,6 @@ package lgp.core.environment.dataset
 
 import com.opencsv.CSVReader
 import lgp.core.environment.ComponentLoaderBuilder
-import lgp.core.environment.config.Configuration
 import lgp.core.modules.ModuleInformation
 import java.io.FileReader
 import java.io.Reader