+++ src/main/kotlin/lgp/core/evolution/Problem.kt
@@ -7,7 +7,7 @@ import lgp.core.environment.config.ConfigurationLoader
 import lgp.core.environment.constants.ConstantLoader
 import lgp.core.environment.operations.OperationLoader
 import lgp.core.evolution.fitness.FitnessFunction
-import lgp.core.evolution.population.EvolutionModel
+import lgp.core.evolution.model.EvolutionModel
 
 data class Description(val description: String)